src/mini_solver_backend/services/restful_api_service/requirements.txt
flask[async]
gunicorn
kink
pydantic
python-dotenv
bcrypt
google-cloud-firestore
google-auth
google-auth-oauthlib
PyJWT
pydantic[email]